+++ src/lib/components/Game.svelte
@@ -5,7 +5,7 @@
 	import { gameStore } from '$lib/stores/gameStore';
 	import ReloadIcon from '$lib/components/icons/ReloadIcon.svelte';
 	import { onMount } from 'svelte';
-	import { getPersistentTarget, setPersistentTarget } from '$lib/utils/storage';
+	import { getPersistentTarget, setPersistentTarget } from '$lib/utils/targetStorage';
 
 	let gameState: GameState;
 	gameStore.subscribe((value) => (gameState = value));
