+++ src/lib/stores/gameStore.ts
@@ -1,12 +1,12 @@
 import { get, writable } from 'svelte/store';
-import type { GameState } from '$lib/interfaces/gameState';
-import { createNewGame } from '$lib/interfaces/gameState';
-import { game } from '$lib/managers/game';
 import { ApplicationState } from '$lib/constants/applicationState';
-import { loadGameState, saveGameState } from '$lib/utils/storage';
+import type { GameState } from '$lib/interfaces/gameState';
+import { game } from '$lib/managers/game';
 import { languageStore } from '$lib/stores/languageStore';
-import { loadCards, loadSingleCard } from '$lib/languages/load';
 import { Tag } from '$lib/constants/tag';
+import { createNewGame } from '$lib/gameState/createNewGame';
+import { saveGameState, loadGameState } from '$lib/gameState/gameStateStorage';
+import { loadCards, loadSingleCard } from '$lib/cards/cardStorage';
 
 /**
  * Creates a writable store to manage the game state and provides several actions
