+++ src/lib/utils/seed.ts
@@ -1,6 +1,26 @@
 import { type Card } from '$lib/interfaces/card';
 
 /**
+ * Shuffles an array of cards in place using a seeded random number generator.
+ * This function ensures a deterministic shuffle based on the provided seed, which allows
+ * reproducibility of the shuffle order.
+ *
+ * @param array - The array of `Card` objects to shuffle.
+ * @param seed - The seed value to initialize the random number generator.
+ *               This ensures the shuffle order is consistent across calls with the same seed.
+ * @returns A new shuffled array of `Card` objects.
+ */
+export function seededShuffle(array: Card[], seed: number): Card[] {
+	const random = seededRandom(seed);
+
+	for (let i = array.length - 1; i > 0; i--) {
+		const j = Math.floor(random() * (i + 1));
+		[array[i], array[j]] = [array[j], array[i]];
+	}
+	return array;
+}
+
+/**
  * Creates a seeded random number generator using a linear congruential generator (LCG).
  * The generator produces pseudorandom numbers based on the given seed.
  *
@@ -24,24 +44,4 @@ function seededRandom(seed: number): () => number {
 		// Normalize the result to a floating-point number in the range [0, 1].
 		return (s - 1) / 2147483646;
 	};
-}
-
-/**
- * Shuffles an array of cards in place using a seeded random number generator.
- * This function ensures a deterministic shuffle based on the provided seed, which allows
- * reproducibility of the shuffle order.
- *
- * @param array - The array of `Card` objects to shuffle.
- * @param seed - The seed value to initialize the random number generator.
- *               This ensures the shuffle order is consistent across calls with the same seed.
- * @returns A new shuffled array of `Card` objects.
- */
-export function seededShuffle(array: Card[], seed: number): Card[] {
-	const random = seededRandom(seed);
-
-	for (let i = array.length - 1; i > 0; i--) {
-		const j = Math.floor(random() * (i + 1));
-		[array[i], array[j]] = [array[j], array[i]];
-	}
-	return array;
 }
